+++ diorite-utils/config/src/main/java/org/diorite/config/annotations/PredefinedComments.java
@@ -22,39 +22,24 @@
  * SOFTWARE.
  */
 
-package org.diorite.config.serialization.annotations;
+package org.diorite.config.annotations;
 
-import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.diorite.config.serialization.DeserializationData;
-import org.diorite.config.serialization.SerializationData;
-
 /**
- * Used to mark class and that should be used for serialization. <br>
- * Serialization method must be matching one of this patterns:
- * <ol>
- * <li>static void nameOfMethod(T, {@link SerializationData})</li>
- * <li>static void nameOfMethod({@link SerializationData}, T)</li>
- * <li>static void nameOfMethod(Object, {@link SerializationData})</li>
- * <li>static void nameOfMethod({@link SerializationData}, Object)</li>
- * <li>void nameOfMethod({@link SerializationData})</li>
- * </ol>
- * Deserialization method must be matching one of this patterns:
- * <ol>
- * <li>static T nameOfMethod({@link DeserializationData})</li>
- * <li>constructor({@link DeserializationData})</li>
- * </ol>
- *
- * @see org.diorite.config.serialization.Serializable
- * @see org.diorite.config.serialization.Serializer
+ * Used to setup comments for class fields.
  */
-@Documented
-@Target({ElementType.METHOD, ElementType.CONSTRUCTOR, ElementType.TYPE})
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Serializable
+public @interface PredefinedComments
 {
+    /**
+     * Predefined comments.
+     *
+     * @return predefined comments.
+     */
+    PredefinedComment[] value();
 }
