+++ diorite-utils/config/src/main/java/org/diorite/config/annotations/SerializableAs.java
@@ -22,24 +22,26 @@
  * SOFTWARE.
  */
 
-package org.diorite.config.serialization.annotations;
+package org.diorite.config.annotations;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Used to add footer to type that will be parsed into {@link org.diorite.config.serialization.comments.DocumentComments}
+ * Allows to serialize given type as different type, like you can implement serialization in implementation class, but use API class as type.
  */
+@Documented
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Footer
+public @interface SerializableAs
 {
     /**
-     * Footer value.
+     * Class to use instead.
      *
-     * @return footer content.
+     * @return class used to registration.
      */
-    String[] value();
+    Class<?> value();
 }
