+++ diorite-utils/config/src/main/java/org/diorite/config/annotations/StringSerializable.java
@@ -22,10 +22,36 @@
  * SOFTWARE.
  */
 
-/**
- * Annotations used by serialization API.
- */
-@NonnullByDefault
-package org.diorite.config.serialization.annotations;
+package org.diorite.config.annotations;
 
-import org.diorite.annotations.NonnullByDefault;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.diorite.config.serialization.StringSerializer;
+
+/**
+ * Used to mark class and methods that should be used for string serialization. <br>
+ * Serialization method must be matching one of this patterns:
+ * <ol>
+ * <li>static String nameOfMethod(T)</li>
+ * <li>static String nameOfMethod(Object)</li>
+ * <li>String nameOfMethod()</li>
+ * </ol>
+ * Deserialization method must be matching one of this patterns:
+ * <ol>
+ * <li>static T nameOfMethod(String)</li>
+ * <li>constructor(String)</li>
+ * </ol>
+ *
+ * @see org.diorite.config.serialization.StringSerializable
+ * @see StringSerializer
+ */
+@Documented
+@Target({ElementType.METHOD, ElementType.CONSTRUCTOR, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface StringSerializable
+{
+}
